dashboard/requirements.txt
dash==2.18.2
plotly==5.24.1
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
dash-bootstrap-components==1.6.0